# cmake/AddCephTest: bind crimson unittest to different cores