@rosinL 2024-01-26
For x86 make check， the longest crimson unittest job `unittest-transaction-manager`, costtime reduced from 1940.81 sec to 983.14 sec. As long tail job `run-rbd-unit-tests-61.sh`, the total test time reduced not so much. As make check arm is still running ,there is no Arm test data for now, but it does same effect on Arm.
Before(x86)
```
287/290 Test #257: unittest-omap-manager .....................   Passed  1642.30 sec
288/290 Test  #32: run-rbd-unit-tests-61.sh ..................   Passed  1728.38 sec
289/290 Test #264: unittest-fltree-onode-manager .............   Passed  1688.28 sec
290/290 Test #251: unittest-transaction-manager ..............   Passed  1940.81 sec

100% tests passed, 0 tests failed out of 290

Total Test time (real) = 1980.95 sec
``` 
After(x86)
```
287/290 Test #251: unittest-transaction-manager ..............   Passed  983.14 sec
288/290 Test  #31: run-rbd-unit-tests-1.sh ...................   Passed  1529.75 sec
289/290 Test #147: check-generated.sh ........................   Passed  1686.12 sec
290/290 Test  #32: run-rbd-unit-tests-61.sh ..................   Passed  1689.13 sec

100% tests passed, 0 tests failed out of 290

Total Test time (real) = 1689.17 sec
```