@athanatos 2024-01-26
@rosinL So different crimson unittest processes are being bound to core 0?  I'd have expected --smp 0 to limit any particular process to a single reactor, but not to limit it to any specific core.