@rosinL 2024-01-27
@athanatos As I test, with --smp 0 , the test will report `runtime error: division by zero`. Are there some mistakes, please let me know.