@rosinL 2024-01-27
I run 2 crimson unittest job parallelly
Without `--cpuset`, all the jobs run on cpu0
```
root@ubuntu:/home# ps -ef |grep unittest
root     3373367 1082838 59 01:27 pts/1    00:00:02 ./build/bin/unittest-fltree-onode-manager --memory 256M --smp 1
root     3373370  251881 56 01:27 pts/2    00:00:02 ./bin/unittest-transaction-manager --memory 256M --smp 1
root     3373374  715758  0 01:27 pts/4    00:00:00 grep --color=auto unittest
root@ubuntu:/home# taskset -acp 3373367
pid 3373367's current affinity list: 0-47
pid 3373368's current affinity list: 0
pid 3373369's current affinity list: 0
root@ubuntu:/home# taskset -acp 3373370
pid 3373370's current affinity list: 0-47
pid 3373371's current affinity list: 0
pid 3373372's current affinity list: 0
``` 
With `--cpuset` to separate jobs to different cores
```
root@ubuntu:/home# ps -ef |grep unittest
root     3373379 1082838 93 01:28 pts/1    00:00:10 ./build/bin/unittest-fltree-onode-manager --memory 256M --smp 1 --cpuset 1
root     3373382  251881 96 01:28 pts/2    00:00:06 ./bin/unittest-transaction-manager --memory 256M --smp 1 --cpuset 0
root     3373386  715758  0 01:28 pts/4    00:00:00 grep --color=auto unittest
root@ubuntu:/home# taskset -acp 3373379
pid 3373379's current affinity list: 0-47
pid 3373380's current affinity list: 1
pid 3373381's current affinity list: 1
root@ubuntu:/home# taskset -acp 3373382
pid 3373382's current affinity list: 0-47
pid 3373383's current affinity list: 0
pid 3373384's current affinity list: 0
```