@rosinL 2024-01-27
@athanatos If there are other better solutions, please let me know, thanks.