@rosinL 2024-01-30
@tchaikov According my understand,  we need `taskset or numactl or seastar arg --cpuset` to distribute crimson tests to different cores. ctest resource allocation can consider CPU shards as a resource, but after CPU shard allocated, ctest cann't bind crimson test to the CPU shard automatically, still need `taskset or numactl or seastar arg --cpuset` to bind cores manually, It seems more  complicated, Am I right?