@rosinL 2024-02-04
> ctest populates the allocated resource using environment variables like CTEST_RESOURCE_GROUP_<num>_<resource-type>. i'd suggest rework the crimson tests to respect these variables, and set the seastar application config accordingly

@tchaikov Do you mean like this:
```
int main(int argc, char** argv) {
    seastar::app_template app;
    // get cpuset from CTEST_RESOURCE_GROUP_<num>_<resource-type>
    // get new_argv with argc + 2  elements
    // copy argv to new_argv
    // new_argv[argc]="--cpuset"
    // new_argv[argc+1]="startcpuid-endcpuid"
    app.run(argc, new_argv, [] {
        return crimson_test_func();
    });
}
```