@tchaikov 2024-02-04
> > ctest populates the allocated resource using environment variables like CTEST_RESOURCE_GROUP__. i'd suggest rework the crimson tests to respect these variables, and set the seastar application config accordingly
> 
> @tchaikov Do you mean like this:
> 
> ```
> int main(int argc, char** argv) {
>     seastar::app_template app;
>     // get cpuset from CTEST_RESOURCE_GROUP_<num>_<resource-type>
>     // get new_argv with argc + 2  elements
>     // copy argv to new_argv
>     // new_argv[argc]="--cpuset"
>     // new_argv[argc+1]="startcpuid-endcpuid"
>     app.run(argc+2, new_argv, [] {
>         return crimson_test_func();
>     });
> }
> ```

something like:

```c++
struct ctest_resource {
  int id;
  int slots;
};

std::vector<ctest_resource> parse_ctest_resources(const std::string resource_spec) {
  vector<string> resources;
  boost::split(resources, resource_spec, line,boost::is_any_of(";"));
  std::regex res_regex("id:([0-9]+),slots:([0-9]+)");
  vector<ctest_resource> ctest_resources;
  for (auto& resource : resources) {
    std::smatch matched;
    std::regex_match(resource, matched, res_regex));
    int id = stoi(matched[0].str());
    int slots = stoi(matched[1].str());
    ctest_resources.emplace_back(id, slots);
  }
  return ctest_resources;
}

std::optional<seastar::resource::cpuset> cpuset_from_ctest_resource_group() {
  int nr_groups = 0;
  if (auto group_count = getenv("CTEST_RESOURCE_GROUP_COUNT"); group_count != nullptr) {
    nr_groups = std::stoi(group_count);
  } else {
    return {};
  }
  seastar::resource::cpuset cpuset;
  for (int num = 0; num < nr_groups; num++) {
    std::string resource_type_name;
    fmt::format_to(std::back_inserter(resource_type_name), "CTEST_RESOURCE_GROUP_{}", num);
    // only a single resource type is supported for now
    std::string resource_type = getenv(resource_type_name.data());
    std::transform(resource_type.begin(), resource_type.end(), resource_type.begin(), ::toupper);
    std::string resource_group;
    fmt::format_to(std::back_inserter(resource_group), "CTEST_RESOURCE_GROUP_{}_{}", num, resource_type);
    std::string resource_spec = getenv(resource_group.data());
    if (group == "CPUS") {
      for (auto& resource : parse_ctest_resources(resource_spec)) {
        // each id has a single cpu slot
        cpuset.insert(resource.id);
      }
    } else {
      fmt::print(cerr, "unsupported resource type: {}", resource_type);
    }
  }
  return cpuset;
}

int main(int argc, char** argv) {
  app_template::seastar_options opts;
  if (auto cpuset = cpuset_from_ctest_resource_group(); cpuset)
    opts.smp_opts.cpuset.set_value(*cpuset);
  }
  app_template app{std::move(opts)};
  return app.run(argc, argv, [] {
    return run_tests();
  };
}
```