@rosinL 2024-02-19
For x86 irvingi07 with 24 cores, the Total Test time keep the same as the long tail job `check-generated.sh`,  the longest crimson unittest `unittest-transaction-manager` job's time decreased from 2522.75 sec to 1375.03 sec, nearly 2X performance improvement.
Before
```
285/290 Test #257: unittest-omap-manager .....................   Passed  2010.54 sec
286/290 Test  #32: run-rbd-unit-tests-61.sh ..................   Passed  2425.26 sec
287/290 Test #148: readable.sh ...............................   Passed  2458.99 sec
288/290 Test #264: unittest-fltree-onode-manager .............   Passed  2093.77 sec
289/290 Test #251: unittest-transaction-manager ..............   Passed  2522.75 sec
290/290 Test #147: check-generated.sh ........................   Passed  2846.91 sec

100% tests passed, 0 tests failed out of 290

Total Test time (real) = 2919.83 sec
```
After
```
285/290 Test  #52: unittest_bufferlist .......................   Passed  1238.60 sec
286/290 Test #251: unittest-transaction-manager ..............   Passed  1375.03 sec
287/290 Test  #31: run-rbd-unit-tests-1.sh ...................   Passed  2063.50 sec
288/290 Test  #32: run-rbd-unit-tests-61.sh ..................   Passed  2481.31 sec
289/290 Test #148: readable.sh ...............................   Passed  2446.87 sec
290/290 Test #147: check-generated.sh ........................   Passed  2838.97 sec

100% tests passed, 0 tests failed out of 290

Total Test time (real) = 2914.18 sec
```